Adjusts wording and removes disk writing requirement.
We should try to avoid using acronyms in specifications as it makes it harder for new participants in the ecosystem to process, so I have changed all references to EL and CL to execution layer and consensus layer.  Also, in many cases in this document the term EL was used when EL client was meant, so I appended "client" as appropriate.

While writing to the filesystem is useful in some setups, in others it may be a security risk.  For this reason, the **MUST** write to disk was changed to a **SHOULD** write to disk so that a client that doesn't write the token to disk isn't out of compliance.