Clients should have autonomy for mechanism of configuration.
This spec, as currently written is incredibly prescriptive about exactly how the key should be provided to the client.  This doesn't allow clients to differentiate in ways that are useful to their customers without falling out of spec compliance.

One can easily imagine a user wanting to configure this via environment variable, or CLI parameter rather than via file.  Also, files on disk may not be secure against reading by unauthorized users (for example, disk backups may not be encrypted) while access to startup parameters and environment variables may be better protected.  A client may also integrate with a secure secret storage system that allows the client to fetch secrets from a server over a secure connection at startup, thus protecting the key from anyone who doesn't have the ability to read the contents of RAM while the client is running.

Note, this change should be merged after #298 as it incorporates some of the wording changes found there.

I can appreciate the value in having consistency in configuration parameter naming, I don't think we should be setting such things as required.  Nethermind, for example, namespaces its configuration variables and having this be un-namespaced is likely to cause more confusion for users and more complexity in their client.