Add `engine_newPayloadHeaderV1`
While `engine_forkchoiceUpdatedV1` specifies that passing an unknown
`headBlockHash` MAY trigger a sync, EL implementations practically
require `engine_newPayloadV1` to be called as well.

This is prohibitive for applications where the CL is a light client
that does not have the capabilities to obtain full beacon block data.
To better support that scenario, a new endpoint is proposed to provide
the EL client software with just the `ExecutionPayloadHeader`, therefore
simplifying integration with existing sync logic.

Furthermore, this unlocks additional experiences:
1. A CL that is syncing optimistically after a prolonged outage can run
   CL light client sync to obtain the latest `ExecutionPayloadHeader`.
   This allows the EL to start syncing to a state close to wall time
   without having to wait for the CL optimistic sync to catch up.
2. EL light client implementations that wish to sync only data relevant
   to a certain use case (e.g., transactions for a specific wallet) can
   use the `ExecutionPayloadHeader` to decide what blocks are relevant.
   The new endpoint enables combined CL/EL deployments.