Sort RPC methods in docs (#301)
The methods that appear in the published docs are not sorted by method
name. This may not matter when looking for a particular method (as you
can use Ctrl-F/Cmd-F in your browser), but not so when using the list as
a whole. In our case, we have a proxy to Ethereum in our project, and we
want to ensure we have proper test coverage for all public Ethereum RPC
methods. This is difficult when the official list of methods is unsorted
because we have to double- and triple-check that we've accounted for all
of them on our side. If the list were sorted then this task would be a
lot easier.