eth: add optional v value for 2930 and 1559 txs (#446)
It looks like many clients are currently either not returning yParity or returning both v and yParity. To align the spec with this, I'm proposing we add an optional (but deprected) v value for 2930 and 1559 txs. This way clients can continue use it and tooling can have some space to upgrade to yParity.

---------

Co-authored-by: Felix Lange <fjl@twurst.com>