feat: add returndata to tx receipts
right now, returndata is not part of the tx receipt. this means that
consumers of the API need to find alternative ways to fetch the
returndata from a transaction, the most common being
`debug_traceTransaction`. however, this is non-standard and not
universally supported. it would be helpful if the returndata were simply
part of the transaction receipt.